refactor: use ngraph only
