+++ packages/path-finding/src/types.ts
@@ -7,7 +7,7 @@ export type PathFinder = (
 
 export interface Graph {
   addNode(v: Vector): void;
-  addLink(from: Vector, to: Vector, weight: number): void;
+  addLink(from: Vector, to: Vector): void;
   hasNode(v: Vector): boolean;
   removeNode(v: Vector): void;
   getLinks(v: Vector): Vector[];
